requirements.txt
llama-index
llama-index-llms-google-genai
llama-index-readers-web
python-dotenv
requests
beautifulsoup4
